new story refact
